# [NCL-5611] Protobuf for LocalFile / KojiBuild