--- a/core/src/main/java/org/jboss/pnc/build/finder/core/LocalFile.java
+++ b/core/src/main/java/org/jboss/pnc/build/finder/core/LocalFile.java
@@ -15,14 +15,6 @@
  */
 package org.jboss.pnc.build.finder.core;
 
-import java.io.IOException;
-import java.io.ObjectInput;
-import java.io.ObjectOutput;
-import java.util.Set;
-
-import org.infinispan.commons.marshall.AdvancedExternalizer;
-import org.infinispan.commons.marshall.SerializeWith;
-import org.infinispan.commons.util.Util;
 import org.infinispan.protostream.annotations.ProtoFactory;
 import org.infinispan.protostream.annotations.ProtoField;
 
@@ -31,8 +23,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 @JsonDeserialize
-// TODO: Remove this annotation once conversion to protobuf is done
-@SerializeWith(LocalFile.LocalFileExternalizer.class)
 public class LocalFile {
     private final String filename;
 
@@ -58,46 +48,5 @@ public class LocalFile {
     @Override
     public String toString() {
         return "LocalFile{" + "filename='" + filename + '\'' + ", size=" + size + '}';
-    }
-
-    // TODO: Remove this class once conversion to protobuf is done
-    public static class LocalFileExternalizer implements AdvancedExternalizer<LocalFile> {
-        private static final long serialVersionUID = -3322870654564600039L;
-
-        private static final Integer ID = (Character.getNumericValue('L') << 16) | (Character.getNumericValue('F') << 8)
-                | Character.getNumericValue('E');
-
-        private static final int VERSION = 1;
-
-        @Override
-        public void writeObject(ObjectOutput output, LocalFile object) throws IOException {
-            output.writeInt(VERSION);
-            output.writeUTF(object.getFilename());
-            output.writeLong(object.getSize());
-        }
-
-        @Override
-        public LocalFile readObject(ObjectInput input) throws IOException, ClassNotFoundException {
-            int version = input.readInt();
-
-            if (version != VERSION) {
-                throw new IOException("Invalid version: " + version);
-            }
-
-            String filename = input.readUTF();
-            long size = input.readLong();
-
-            return new LocalFile(filename, size);
-        }
-
-        @Override
-        public Set<Class<? extends LocalFile>> getTypeClasses() {
-            return Util.asSet(LocalFile.class);
-        }
-
-        @Override
-        public Integer getId() {
-            return ID;
-        }
     }
 }