This commit adds support for Protobuf serialization to Infinispan.

The change to use Protobuf is motivated by:

1. JBoss marshalling is deprecated in Infinispan
2. Use of Protobuf is required to use Infinispan remote servers

Implementation
--------------

`KojiBuild` and `KojiArchiveInfo` have Protobuf adapters that marshall
the objects into JSON and vice-versa. Their specific serializer code is
removed as a result.

`ListKojiArchiveInfoProtobufWrapper` object is created to wrap around a
list of `KojiArchiveInfo` objects. This is done to workaround an
Infinispan bug (https://issues.redhat.com/browse/IPROTO-219)

Finally, the configuration for the global cache removes the use of the
specific serializers for `LocalFile` and `KojiBuild` and are instead
replaced by the `ProtobufSerializerImpl` which contains the Protobuf
definitions.

Benchmark results
-----------------

The results also depend on the Koji server response and shouldn't be
taken too seriously. An outlier would indicate serializing might take
more time. From the results, they look pretty similar.

|Without cache: JBoss Marshall|Without cache: Protobuf + JSON|
|----------------------------:|-----------------------------:|
|11m24s|3m38s|
|3m33s|4m30s|
|5m14s|7m06s|

This tests how fast we can retrieve data from Infinispan and marshall
them to a Java Object. From the results, both JBoss Marshalling and
Protobuf+JSON take about the same time

|With cache: JBoss Marshall|With cache: Protobuf + JSON|
|----------------------------:|-----------------------------:|
|5.7s|4.9s|
|4.9s|4.8s|
|4.8s|4.8s|

The benchmark results indicate that JBoss Marshalling takes about the
same time as using Protobuf to marshall and unmarshall data from
Infinispan