@dwalluck 2021-08-17
Test failures now and can we do something about the unchecked warnings?

We've got deprecated warnings from org.jboss.pnc.constants, but I don't know who should be responsible for that code. It's not from your PR, but it's from PNC.